Lots of configuration loader additions.  Added attribute to specify the source an object can be decoded from - ConfigTypes.Value or ConfigTypes.Node.
Templates are successfully cloned from existing bodies, and properties class can configure celestial body parameters short of biomes and science values.

Body can successfully bring up a PSystemBody from a template and do initial creation for bare bring ups.

ScaledPlanetRimAerial now inherits from ScaledPlanetSimple as the prior is a logical extension of the latter.  This allows ScaledPlanetRimAerial to be used in place of ScaledPlanetSimple in scaled version config, but then be able to present the expanded functionality to atmospheric configuration.

Added IEnumerable<T> supporting classes for enumerating the parents of celestial bodies and game objects.  In the case of a potential star systems merge, this allows for predicates and Linq to be used to search for stars.

Signed-off-by: Nathaniel Lewis <linux.robotdude@gmail.com>