I just realized that a lot of the stuff I added for Kittopia is useless
An edit-constructor isn't even needed, since the existing values are added
to the respective list automatically by Kopernicus. Everything Kittopia needs
to do is to add and remove elements from the list. Having a constructor for editing
objects, or passing low level API (i.e. KSP) objects to it is against the purpose
of Kittopia using the Kopernicus API.