Tell the stock scatter spawner to only spawn low-poly cubes
This prevents issues with having a high-poly basemesh, because the
stock code will try to create a gargantuan mesh from it that exceeds
any vertex limit unity has. Since the stock scatters are never visible
anyways, just tell it to build a useless mesh.

Also, this replaces the obj importer with a different version (from
cineboxandrew, licensed under Apache2), because the previous one tried to
create a 100000 vert mesh when loading one with 20000 verts.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>