Fixed exception in ApplyFlagFixes by removing the static from the following: 	private  void ApplyFlagFixes() 	private  void FixFlags(DestructibleBuilding data) 	private  void FixFlags(Upgradeables.UpgradeableFacility data0, int data1) 	private  void FixFlags()
The static isn't needed because the entire class is always there because of the DontDestroyOnLoad()

Also, commented out the body.orbitDriver.UpdateOrbit(); at line 350.
This was causing the hang at the main menu with New Horizons.
I found that using QuickStart to bypass the main menu was letting NH
work normally.