Force KSP to use the highest terrain quality available
We need to do this to ensure that the stock system stays the same
consistently, across installations. If the surface material can change
at every time, noone can safely edit it anymore, because the new values
might not even be read.

As an alternative to the ingame switcher, the default material configs
can be dumped and installed as Kopernicus mods in GameData

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>