Refactor FloatCurve loading
A float curve is hard to express with Kopernicus config loader because
it is a simple value that is loaded through a node. Previously it was
done with a "loader in a loader", but that violates the assumption that
Kittopia makes that setting the loader will also update the value
backing it - which is not the case if you wrap the loader with another
one. This resulted in float curve editing being broken in Kittopia.

This new way doesn't look as nice as the old one does, but it works the
same, and makes it possible to properly edit the curves in Kittopia.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>