Improve the TextureAtlas disabling.
The PQS ignores modEnabled = false if it was set after the PQS has
started. This means that we have to figure out if any of the mods was
active, and if yes, tell the PQS to rescan them. We cannot rescan them
unconditionally because that would invoke OnSetup again and enter an
infinite loop.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>