Simplify the names for surface materials
The current names are way too long and complex, which makes them hard to
use in any conversation (like when searching for a bug). It's also
misleading that all of these are prefixed with Atmospheric, because
there shouldn't be anything stopping you from using them on Vaccum
bodies.

This commit simplifies the names a bit, while keeping around the old
identifiers so that existing configs will continue to work like before.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>