Support Breaking Ground anomalies out of the box
The surface anomalies that were added with the Breaking Ground DLC
have their own stock configuration system, but they rely on a PQSMod
that needs to be added to a body with anomalies.

That in itself wouldn't be a huge problem, because you could always
template bodies that already have this mod. However, the PQSMod has a
design flaw: It doesn't actively look for the body it is attached to,
instead it simply stores the name in a string variable.

That means if you create a body called "Kopernicus", with Vall as it's
template, the string variable will still contain the string "Vall", and
the ROC system will spawn the anomalies for Vall on that body. If you
add anomalies for "Kopernicus", they will by ignored.

This code does the same that the "MyRocksAreBiggerThanYours" plugin that
I wrote for JNSQ a while ago does, but that was always just a stopgap
solution. Integrating it into Kopernicus is more universal.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>