perf: Move actual work of CompileRGB to a job
The single most expensive operation in this method is
Texture2D.GetRawTextureData. We can launch a job to do the actual
conversion in parallel to getting the raw texture data.

For a stock + parallax continued save this improves scene switch times
by 70ms. For Sol this improves scene switch times by another 700ms.