Make missing OnDemand textures a load-time error
With some experimence debugging issues for Sol this seems to be a pretty
big footgun. The most common visible issue when OnDemand textures fail
to load is getting stuck in a loading screen. I think it would be much
better if users saw the regular "kopernicus failed to load a planet"
popup in the main menu, instead of getting stuck in an infinite loading
screen.

This changes the MapSO parsers for MapSODemand to throw an exception if
during parsing if the texture doesn't exist. This doesn't catch all the
errors, but I expect other errors to be far less common than textures
not being found due to a broken install.