Use burst to accelerate KopernicusCBAttributeMapSO creation
This is a little cheeky, but it turns out that you can just copy the
BurstCompile attribute into your project and then use it to accelerate
your jobs. Unity itself will take care of calling the correct job
functions if the appropriate library is present, otherwise will just run
the C# code as it normally would.

This uses that trick to burst-compile the compilation job for
KopernicusCBAttributeMapSO.