Make MapSODemand more resilient against errors
This introduces a third error state that basically just makes the map
always return a default value. This should be better than the
alternative, where it would either try to repeatedly load the texture or
just throw a NRE and abort.

Actually running into this situation should be pretty rare now that we
error upon finding missing OnDemand textures, but that's no reason not
to fail more gracefully.